Added functionality
1. Added in functionality to take the meteorological forecast data time
series from a different data frame (fc_data) to that used to drive
tamsat-alert. fc_data is a data frame containing time series of all of
the driving variables for tamsat-alert. fc_data is an argument to the
tamsat_alert function.

2. Added in functionality to select either precipitation or temperature
as the meteorological forecast variable. This is accomplished by a new
user defined variable: met_ts_varname, which can be set to either
'precipitation' or 'temperature'. The dataframe temperature and
precipitation column names are now also set in the arguments to
tamsat-alert as precipitation_rate_str and temperature_str, which are
defaulted to the back end defined pandas dataframe column names in an
analogous manner to tamsat_alert_sm.py